Add Xcode project and update to use Accelerate in Xcode.
This commit adds an Xcode project. Since Accelerate comes for free in Xcode, it makes sense to use this instead of forcing the user to install LAPACKE/OpenBLAS as on Linux. Unfortunately Accelerate has clapack, not LAPACKE, which is a slightly different interface. The NIFTY_XCODE_BUILD flag is used to determine when being built with Xcode, in which case Accelerate is used and the functions are called accordingly.

NOTE: There is some bug in how I'm calling CBLAS functions... the transpose called from the SVD test fails in Xcode, for example.