Add trig function overloads and tests
Added in overloads for trig functions of vectors, matrices, and tensors.

Basic tests just compare a few results to caluculator/matlab output, but it's not likely that we need much more testing because all the heavy lifting is done by standard library functions.

Changed the way that the standard library functions are imported. Before, we declared e.g. cos as a global variable that pointed to Glibc.cos. The problem with that is it shows up as a global variable not a function. If we create a function called cos that just calls Glibc.cos, that's fine until somebody imports Nifty and Glibc in the same file, which causes it to be ambiguous. Now, we use the @_exported attribute on the import statement to just expose the symbols in Nifty. This solves the type and ambiguity problems, but it's currently an undocumented feature... maybe it goes away.