Remove use of Dispatch module
The Travis CI build started breaking on the last commit with vague error messages. I think it might be related to use of the Dispatch module. This commit removes use of Dispatch, and all dependent functionality, in an attempt to narrow down the cause of breakage.

Note: previous commit that breaks on Travis builds just fine on my local mac machine.