Update data frame to use latest data series
Previously, the data frame maintained an index list and a list of lists containing the various columns. This would have required the data frame to implement all the filling, inserting, etc by itself, duplicating what's done by the series. This commit redoes the frame to contain a list of series, and just enforces that all series have the same indexes. This allows for deferring much of the code to the data series objects.

In addition to the data frame rework, tests were add and minor bugs fixed in the series and interp1 implementations.