Checkpoint - Increase efficiency and extend API.
This commit increases efficiency and speed mostly by improving naive implementations (e.g. binary search instead of linear). It also adds some functionality. However, efficiency can still be improved and the API is inconsistent and in places awkward. This commit is just one more iteration in the refinement process and, though it passes the basic test cases, should not be seen as a final product.